spline #810 bugfixes
