+++ persistence/src/main/scala/za/co/absa/spline/persistence/ArangoDatabaseFacade.scala
@@ -70,7 +70,7 @@ object ArangoDatabaseFacade {
 
   private def warmUpDb(db: ArangoDatabaseAsync): Unit = {
     withWorkaroundForArangoAsyncBug {
-      db.getInfo.get()
+      db.exists.get
     }
   }
