+++ persistence/src/main/scala/za/co/absa/spline/persistence/DatabaseVersionManager.scala
@@ -20,7 +20,6 @@ import com.arangodb.async.ArangoDatabaseAsync
 import za.co.absa.commons.version.Version
 import za.co.absa.commons.version.Version._
 import za.co.absa.commons.version.impl.SemVer20Impl.SemanticVersion
-import za.co.absa.spline.persistence.ArangoImplicits.ArangoDatabaseAsyncScalaWrapper
 import za.co.absa.spline.persistence.DatabaseVersionManager._
 import za.co.absa.spline.persistence.model.DBVersion.Status
 import za.co.absa.spline.persistence.model.NodeDef.DBVersion
@@ -29,6 +28,9 @@ import scala.compat.java8.FutureConverters.CompletionStageOps
 import scala.concurrent.{ExecutionContext, Future}
 
 class DatabaseVersionManager(db: ArangoDatabaseAsync)(implicit ec: ExecutionContext) {
+
+  import ArangoImplicits._
+
   def insertDbVersion(currentVersion: SemanticVersion): Future[SemanticVersion] = {
     val dbVersion = model.DBVersion(currentVersion.asString, model.DBVersion.Status.Current)
     for {
@@ -52,7 +54,7 @@ class DatabaseVersionManager(db: ArangoDatabaseAsync)(implicit ec: ExecutionCont
   private def getDBVersion(status: model.DBVersion.Status.Type) = {
     db
       .collection(DBVersion.name)
-      .exists().toScala
+      .existsCollection()
       .flatMap(exists =>
         if (exists) db.queryOptional[String](
           s"""
